requirements.txt.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.1.2
django-bootstrap==0.2.4
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-embed-video==1.3.3
django-js-asset==1.2.2
django-taggit==1.3.0
dnspython==2.0.0
email-validator==1.1.1
filelock==3.0.12
idna==2.10
importlib-metadata==2.0.0
Pillow==8.0.0
pipenv==2020.8.13
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
validate-email==1.3
virtualenv==20.1.0
virtualenv-clone==0.5.4
zipp==3.3.2
